--- a/include/openmc/random_ray/flat_source_domain.h
+++ b/include/openmc/random_ray/flat_source_domain.h
@@ -112,6 +112,7 @@ public:
   virtual void flux_swap();
   virtual double evaluate_flux_at_point(Position r, int64_t sr, int g) const;
   double compute_fixed_source_normalization_factor() const;
+  void flatten_xs();
 
   //----------------------------------------------------------------------------
   // Static Data members
@@ -144,6 +145,18 @@ public:
   vector<float> scalar_flux_new_;
   vector<float> source_;
   vector<float> external_source_;
+
+  // 2D arrays stored in 1D representing values for all materials x energy
+  // groups
+  int n_materials_;
+  vector<float> sigma_t_;
+  vector<float> nu_sigma_f_;
+  vector<float> sigma_f_;
+  vector<float> chi_;
+
+  // 3D arrays stored in 1D representing values for all materials x energy
+  // groups x energy groups
+  vector<float> sigma_s_;
 
 protected:
   //----------------------------------------------------------------------------