--- a/include/openmc/random_ray/random_ray_simulation.h
+++ b/include/openmc/random_ray/random_ray_simulation.h
@@ -30,12 +30,13 @@ public:
   void print_results_random_ray(uint64_t total_geometric_intersections,
     double avg_miss_rate, int negroups, int64_t n_source_regions,
     int64_t n_external_source_regions) const;
-  void reset();
-  // Contains all flat source region data
-  unique_ptr<FlatSourceDomain> domain_;
 
   //----------------------------------------------------------------------------
   // Data members
+  
+  // Contains all flat source region data
+  unique_ptr<FlatSourceDomain> domain_;
+
 private:
   // Random ray eigenvalue
   double k_eff_ {1.0};