--- a/include/openmc/random_ray/random_ray_simulation.h
+++ b/include/openmc/random_ray/random_ray_simulation.h
@@ -33,7 +33,7 @@ public:
 
   //----------------------------------------------------------------------------
   // Data members
-  
+
   // Contains all flat source region data
   unique_ptr<FlatSourceDomain> domain_;