--- a/include/openmc/random_ray/random_ray_simulation.h
+++ b/include/openmc/random_ray/random_ray_simulation.h
@@ -32,12 +32,16 @@ public:
     int64_t n_external_source_regions) const;
 
   //----------------------------------------------------------------------------
+  // Accessors
+  FlatSourceDomain* domain() const { return domain_.get(); }
+
+private:
+  //----------------------------------------------------------------------------
   // Data members
 
   // Contains all flat source region data
   unique_ptr<FlatSourceDomain> domain_;
 
-private:
   // Random ray eigenvalue
   double k_eff_ {1.0};