--- a/openmc/settings.py
+++ b/openmc/settings.py
@@ -168,6 +168,9 @@ class Settings:
             cm/cm^3. When disabled, flux tallies will be reported in units
             of cm (i.e., total distance traveled by neutrons in the spatial
             tally region).
+        :adjoint:
+            Whether to run the random ray solver in adjoint mode (bool). The
+            default is 'False'.
 
         .. versionadded:: 0.15.0
     resonance_scattering : dict
@@ -1107,6 +1110,8 @@ class Settings:
                                ('flat', 'linear', 'linear_xy'))
             elif key == 'volume_normalized_flux_tallies':
                 cv.check_type('volume normalized flux tallies', random_ray[key], bool)
+            elif key == 'adjoint':
+                cv.check_type('adjoint', random_ray[key], bool)
             else:
                 raise ValueError(f'Unable to set random ray to "{key}" which is '
                                  'unsupported by OpenMC')
@@ -1906,6 +1911,10 @@ class Settings:
                     self.random_ray['source_shape'] = child.text
                 elif child.tag == 'volume_normalized_flux_tallies':
                     self.random_ray['volume_normalized_flux_tallies'] = (
+                        child.text in ('true', '1')
+                    )
+                elif child.tag == 'adjoint':
+                    self.random_ray['adjoint'] = (
                         child.text in ('true', '1')
                     )