--- a/src/mgxs_interface.cpp
+++ b/src/mgxs_interface.cpp
@@ -146,7 +146,7 @@ void MgxsInterface::create_macro_xs()
         num_energy_groups_, num_delayed_groups_);
     } else {
       // Preserve the ordering of materials by including a blank entry
-      macro_xs_.emplace_back();
+      macro_xs_.emplace_back(false);
     }
   }
 }