… docs/source/methods/random_ray.rst
… docs/source/usersguide/random_ray.rst
… src/random_ray/random_ray.cpp
… src/random_ray/random_ray_simulation.cpp
… src/settings.cpp
… tests/regression_tests/random_ray_adjoint_fixed_source/inputs_true.dat
… tests/regression_tests/random_ray_adjoint_fixed_source/results_true.dat
… tests/regression_tests/random_ray_adjoint_fixed_source/test.py
… tests/regression_tests/random_ray_adjoint_k_eff/inputs_true.dat
… tests/regression_tests/random_ray_adjoint_k_eff/results_true.dat
… tests/regression_tests/random_ray_adjoint_k_eff/test.py