<!--
If you are a first-time contributor to OpenMC, please have a look at our
contributing guidelines:
https://github.com/openmc-dev/openmc/blob/develop/CONTRIBUTING.md
-->

# Description

This PR adds an adjoint flux mode to the random ray solver. When enabled via:

```python
settings.random_ray['adjoint'] = True
```

all results (k-eff, tallies.out, statepoint, etc) will be given in terms of the adjoint flux. Closes #2835, closes #2831.

## Details

More details on the physics and how to use/interpret the results are given in new sections of the docs in the random ray theory and random ray user guide areas. A few specific notes are given here:

- The adjoint solve is easily implemented by simply transposing the scattering matrix and swapping the nu_Sigma_f and Chi vectors.

- For the purposes of making the code faster (by improving the MGXS access pattern efficiency) and for making the above transposition/swapping operations trivial, I have created an alternative MGXS data structure native for the random ray solver (as proposed in Issue #2831). This structure is populated based on the existing MGXS object (which is not significantly affected by this change) and is initialized before beginning a random ray solve. The new random ray MGXS data structures are now stored in the `FlatSourceDomain` object as:

```C++
  // 2D arrays stored in 1D representing values for all materials x energy
  // groups
  int n_materials_;
  vector<double> sigma_t_;
  vector<double> nu_sigma_f_;
  vector<double> sigma_f_;
  vector<double> chi_;

  // 3D arrays stored in 1D representing values for all materials x energy
  // groups x energy groups
  vector<double> sigma_s_;
```

The only change to the MGXS object is a minor addition of an `exists_in_model` parameter. The previous MGXS object functioned by allowing for empty material indices to be stored, but did not mark them in any particular manner. When converting MGXS to the flattened random ray format, we don't want to accidentally read data that doesn't exist, so we utilize this flag for skipping these data sets and initializing data to zero instead.

- When the code is run in adjoint mode, a forward flux solve is run first before a subsequent adjoint solve is run. This serves two purposes. The first is so that (internally at least) the forward flux and the adjoint flux vectors are both present in memory at the conclusion of the run -- allowing for easy future implementation of FW-CADIS for weight window generation, or for use with perturbation methods etc. Additionally, when a fixed external source is present, the forward solve is required for computing the external source for the adjoint, which is computed as the inverse of the forward flux. This selection of adjoint source corresponds to a global response, rather than the response function for some specific detector region, making the forward and adjoint pair useful for global variance reduction rather than localized variance reduction uses.

- As two full solves (one forward, one adjoint) are performed, I implemented this by simply destroying the full RandomRaySimulation object after the first solve and then generating a new one for the adjoint. A new function is added for computing the external adjoint source given some flux vector (passed in from the forward solve). There are likely other ways of implementing the adjoint solve, but deleting the first solve's object and then making a new one seemed much easier than manually resetting all the data structures and parameters that the random ray solver touches both internally and throughout OpenMC.

- Tallies etc are not labeled as being specifically adjoint tallies -- they are all implicitly computed with the adjoint flux when the code is run in adjoint mode. It is up to the user to interpret any requested tally scores based on the adjoint.

## Speedup from XS Storage Flattening

On a 69 group 3D problem with about 22 million flat source regions, the runtime data on a dual socket Xeon Platinum 8468 (96 core) node is:

Original:

```
 Total simulation time             = 9.2321e+02 seconds
   Transport sweep only            = 3.1614e+02 seconds
   Source update only              = 5.2710e+02 seconds

```

This PR (flattened XS data):

```
 Total simulation time             = 4.2405e+02 seconds
   Transport sweep only            = 2.9985e+02 seconds
   Source update only              = 4.3163e+01 seconds
```

As can be seen, the flattened XS data provides a modest 5% speedup for the transport stage, but a massive 12x speedup for the source update stage (due to the vastly improved scattering matrix access pattern), resulting in an overall >2.2x speedup.

## Validation

Validation of adjoint solvers is tricky, as most papers presenting new adjoint solvers compare against existing (closed source) adjoint solvers. Worse, the adjoint solutions themselves are typically not presented. As such, there doesn't appear to be any existing benchmark problem for adjoint fluxes that can be used to validate this implementation. In light of this, I present two ways of validating the results, depending on if solving in eigenvalue mode or fixed source mode.

### Eigenvalue validation

Mathematically the adjoint k-eff is statistically the same as the forward k-eff, despite the forward and adjoint fluxes having very different spatial distributions. Thus, one way of verifying that the adjoint mode is working is ensuring that the k-eff values are similar. To do this, I ran 2D C5G7 with linear sources. The forward k-eff is 1.18664 +/- 0.00026, while the adjoint k-eff is 1.18682 +/- 0.00007, which is well within the uncertainty of the forward solve. It is interesting that the adjoint solve has a much lower variance than the forward one. This appears to be caused by the much smoother profile of the adjoint flux, making it easier to integrate.

While the k-eff matches, this could also be caused by a failure mode where setting the adjoint mode had not effect. To ensure the adjoint mode is actually running, we can look at the flux plots to make sure the distributions look different. 

Here's the 2D C5G7 forward flux:

![c5g7_thermal_forward](https://github.com/user-attachments/assets/94265cc7-bf99-44df-8014-2c251b53fb5b)

and adjoint flux:

![c5g7_thermal_adjoint](https://github.com/user-attachments/assets/0bf3cac1-4ffc-4c91-b0e0-fc68f69f64eb)

### Fixed Source Validation

Absent an existing benchmark problem for adjoint fluxes, the best I can do is comparing to a very simple 1 group infinite medium problem. This is implemented as a reflective 10x10x10 cm cube with the following material parameters:

Sigma_t = 0.2
Sigma_a = 0.08
Sigma_s = 0.12
Q = 2 [particles/s] (strength spread uniformly over entire cube)

Analytical phi = Q/Sigma_a = 2 / 0.08 = 25 [cm]
Analytical phi normalized = 25/(10*10*10) = 0.025 [cm/cm^3]

Analytical Q_adj = 1/phi = 1/0.025 = 40 [cm^-3]
Analytical phi_adj = Q_adj/Sigma_a = 40/0.08 = 500 [cm/cm^3]

With OpenMC, I'm able to get the correct forward flux value with multigroup Monte Carlo (MGMC) (i.e., 25, as by default MGMC flux tallies in OpenMC are just tracklength). I'm also able to get the correct forward flux values of 25 and 0.025 with volume normalization off and on with the random ray solver. Finally, I can get the expected adjoint value of 500 with random ray. 

Overall, this study at least validates that the external adjoint source is being calculated via the inverse forward flux as expected, and that the results for an infinite medium 1 group problem are correct at least. I also experimented with making the cube have vacuum boundary conditions (such that the flux now has a spatial shape and a source region mesh is required). I don't have an analytical solution for this alternative problem, but the external adjoint source at least does appear to be getting computed in the expected spatially varying manner given the inverse forward flux.

The full python code for the infinite medium problem is given below:

```python
import openmc
import openmc.mgxs
import numpy as np

def fill_3d_list(n, val):
    return [[[val for _ in range(n)] for _ in range(n)] for _ in range(n)]

###############################################################################
# General Settings
###############################################################################

# Instantiate a Settings, set all runtime parameters, and export to XML
settings_file = openmc.Settings()
settings_file.energy_mode = "multi-group"
settings_file.batches = 200
settings_file.inactive = 100
settings_file.particles = 1000
settings_file.run_mode = 'fixed source'
settings_file.output = {'tallies': True, 'summary': False}

# Create an initial uniform spatial source distribution for sampling rays.
# Note that this must be uniform in space and angle.
width = 10.0
half = width / 2.0
lower_left = (-half, -half, -half)
upper_right = (half, half, half)
uniform_dist = openmc.stats.Box(lower_left, upper_right)
settings_file.random_ray['ray_source'] = openmc.IndependentSource(space=uniform_dist)
settings_file.random_ray['distance_inactive'] = 20.0
settings_file.random_ray['distance_active'] = 200.0
settings_file.random_ray['source_shape'] = 'flat'
settings_file.random_ray['adjoint'] = True
settings_file.random_ray['volume_normalized_flux_tallies'] = True

# Boundary condition (reflective is equivalent to being an infinite homogenous medium)
bc = 'reflective'
#bc = 'vacuum'

###############################################################################
# MGXS
###############################################################################

# Instantiate the energy group data
groups = openmc.mgxs.EnergyGroups([0.0, 1.0])
print(groups)
print(groups.group_edges)

xsdata = openmc.XSdata('infinite material', groups)
xsdata.order = 0

sigma_s = 0.12
scatter_matrix = \
    [[[sigma_s]]]

sigma_t = 0.2
sigma_a = 0.08

xsdata.set_total(np.array([sigma_t]))

xsdata.set_absorption(np.array([sigma_a]))

scatter_matrix = np.array(scatter_matrix)
xsdata.set_scatter_matrix(scatter_matrix)

xsdata.set_fission(np.array([0.0]))
xsdata.set_nu_fission(np.array([0.0]))

xsdata.set_chi(np.array([0.0]))

mg_cross_sections_file = openmc.MGXSLibrary(groups)
mg_cross_sections_file.add_xsdatas([xsdata])
mg_cross_sections_file.export_to_hdf5()

###############################################################################
# Materials
###############################################################################

data = openmc.Macroscopic('infinite material')

# Instantiate some Materials and register the appropriate Macroscopic objects
mat = openmc.Material(name='infinite material')
mat.set_density('macro', 1.0)
mat.add_macroscopic(data)

# Instantiate a Materials collection and export to XML
materials_file = openmc.Materials([mat])
materials_file.cross_sections = "mgxs.h5"

###############################################################################
# Particle Source
###############################################################################

# Create the neutron source
strengths = [1.0]
midpoints = [0.5]
energy_distribution = openmc.stats.Discrete(x=midpoints,p=strengths)
source = openmc.IndependentSource(energy=energy_distribution, space=uniform_dist, constraints={'domains':[mat]}, strength=2.0)

settings_file.source = [source]

###############################################################################
# Base Universe
###############################################################################

# Instantiate a Geometry, register the root Universe, and export to XML
cell = openmc.Cell(fill=mat, name='infinite source region')
cell_uni = openmc.Universe()
cell_uni.add_cells([cell])
    
dim = 10
lattice = openmc.RectLattice()
lattice.lower_left = [-half, -half, -half]
lattice.pitch = [width/dim, width/dim, width/dim]
lattice.universes = fill_3d_list(dim, cell_uni)

lattice_cell = openmc.Cell(fill=lattice, name='lattice')

lattice_uni = openmc.Universe()
lattice_uni.add_cells([lattice_cell])


x_low  = openmc.XPlane(x0=-half,boundary_type=bc) 
x_high = openmc.XPlane(x0=half,boundary_type=bc) 

y_low  = openmc.YPlane(y0=-half,boundary_type=bc) 
y_high = openmc.YPlane(y0=half,boundary_type=bc) 

z_low  = openmc.ZPlane(z0=-half,boundary_type=bc) 
z_high = openmc.ZPlane(z0=half,boundary_type=bc) 

full_domain = openmc.Cell(fill=lattice_uni, region=+x_low & -x_high & +y_low & -y_high & +z_low & -z_high, name='full domain')

root = openmc.Universe(name='root universe')
root.add_cell(full_domain)

geometry = openmc.Geometry(root)

###############################################################################
# Tallies
###############################################################################

estimator = 'tracklength'

mat_filter = openmc.MaterialFilter(mat)
tally = openmc.Tally(name="Infinite Medium Tally")
tally.scores = ['flux']
tally.estimator = estimator
tally.filters = [mat_filter]

# Instantiate a Tallies collection and export to XML
tallies = openmc.Tallies([tally])

###############################################################################
# Model
###############################################################################

model = openmc.model.Model()
model.geometry = geometry
model.materials = materials_file
model.settings = settings_file
model.xs_data = mg_cross_sections_file
model.tallies = tallies

model.export_to_model_xml()
```

## Future Work

- Add a way for the user to provide a "detector" for generating a response function, rather than the global response as it is now. Perhaps this could be done by specifying a specific tally ID to use as the detector?

- Implementation of FW-CADIS should be quite natural now with both the forward and adjoint solutions available on the full source region mesh at program conclusion.



# Checklist

- [x] I have performed a self-review of my own code
- [x] I have run [clang-format](https://docs.openmc.org/en/latest/devguide/styleguide.html#automatic-formatting) (version 15) on any C++ source files (if applicable)
- [x] I have followed the [style guidelines](https://docs.openmc.org/en/latest/devguide/styleguide.html#python) for Python source files (if applicable)
- [x] I have made corresponding changes to the documentation (if applicable)
- [x] I have added tests that prove my fix is effective or that my feature works (if applicable)
<!--
While tests will automatically be checked by CI, it is good practice to
ensure that they pass locally first. See instructions here:
https://docs.openmc.org/en/latest/devguide/tests.html
-->