@jtramm 2024-11-13
Thanks very much for the review @yardasol! I have incorporated your comments on the loop index naming for better consistency. I wasn't able to accept them directly for some reason (github was complaining about trying to commit changes on deleted code?) but it was easy to do the refactoring manually again.